pkg: move gcpsecret to a separate package
It simplifies the dependency tree and fixes a build error for the
send-test-email container.
